+++ openghg_inversions/hbmcmc/inversion_pymc.py
@@ -142,6 +142,7 @@ def inferpymc(
     use_bc: bool = True,
     reparameterise_log_normal: bool = False,
     pollution_events_from_obs: bool = False,
+    no_model_error: bool = False,
 ):
     """
     Uses PyMC module for Bayesian inference for emissions field, boundary
@@ -187,6 +188,9 @@ def inferpymc(
         Add an offset (intercept) to all sites but the first in the site list. Default False.
       verbose:
         When True, prints progress bar
+      no_model_error:
+        When True, only use observation error in likelihood function (omitting min. model error and model error
+        from scaling pollution events.)
 
     Returns:
       outs (array):
@@ -285,7 +289,12 @@ def inferpymc(
             pollution_event = np.abs(pt.dot(hx, x))
 
         pollution_event_scaled_error = pollution_event * sig[sites, sigma_freq_index]
-        epsilon = pt.maximum(pt.sqrt(error**2 + pollution_event_scaled_error**2), min_error)
+
+        if no_model_error is True:
+            epsilon = np.abs(error)
+        else:
+            epsilon = pt.maximum(pt.sqrt(error**2 + pollution_event_scaled_error**2), min_error)
+
         y = pm.Normal("y", mu=mu, sigma=epsilon, observed=Y, shape=ny)
 
         step1 = pm.NUTS(vars=step1_vars)
@@ -641,6 +650,15 @@ def inferpymc_postprocessouts(
         else:
             emds = fp_data[".flux"][emissions_name[0]]
             flux_array_all = emds.data.flux.values
+
+    # HACK: assume that smallest flux dim is time, then re-order flux so that
+    # time is the last coordinate
+    flux_dim_shape = flux_array_all.shape
+    flux_dim_positions = range(len(flux_dim_shape))
+    smallest_dim_position = min(list(zip(flux_dim_positions, flux_dim_shape)), key=(lambda x: x[1]))[0]
+
+    flux_array_all = np.moveaxis(flux_array_all, smallest_dim_position, -1)
+    # end HACK
 
     if flux_array_all.shape[2] == 1:
         print("\nAssuming flux prior is annual and extracting first index of flux array.")